fix: js/css are minified when both minify.js and minify.css are false
